+++ CHANGES.md
@@ -11,6 +11,7 @@ Version 2.1.5 (Pending)
 - Change Staff positions from just role title to team + role title
 - Differentiate between guest and staff for badge text outline
 - Add VoucherType for vouchers (OCC/Hyatt voucher types)
+- Add Years of Service check to Staff check in
 - 2024 Staff/Guest badges
 
 Version 2.1.4 (11/17/2023)